secretary/requirements.txt
python-telegram-bot==21.0
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
apscheduler
requests
